Made the code more OOP
