--- src/main/java/InstructionQueue.java
@@ -1,5 +1,0 @@
-import java.util.ArrayList;
-
-public class InstructionQueue {
-    private ArrayList<Instructions> queue = new ArrayList<>();
-}
