--- src/main/java/Instructions.java
@@ -1,7 +1,0 @@
-public abstract class Instructions {
-    private int instructionIds;
-
-    private String instructionStorage;
-
-    public abstract void implement(User user);
-}
